requirements.txt
dash==2.14.2
plotly==5.17.0
pandas==2.1.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23 